requirements.txt
absl-py==0.1.10
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.5.36
botocore==1.8.50
bs4==0.0.1
bz2file==0.98
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
configparser==3.5.0
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
cycler==0.10.0
DBUtils==1.2
Django==2.0.2
docutils==0.14
et-xmlfile==1.0.1
Flask==0.12.2
Flask-SQLAlchemy==2.1
gensim==3.3.0
html5lib==0.9999999
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
jdcal==1.3
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.2
Keras==2.1.4
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
Naked==0.1.31
neo4j-driver==1.5.3
nltk==3.2.5
numpy==1.14.1
openpyxl==2.5.0
pandas==0.22.0
parsel==1.4.0
passlib==1.7.1
pbr==3.1.1
Pillow==5.0.0
Pinyin2Hanzi==0.1.1
protobuf==3.5.1
py2neo==3.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodome==3.5.1
PyDispatcher==2.0.5
PyMySQL==0.8.0
PyNLPIR==0.5.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypinyin==0.30.0
pysolr==3.7.0
python-dateutil==2.6.1
python-docx==0.8.6
python-Levenshtein==0.12.0
pytz==2018.3
pywin32==221
PyYAML==3.12
queuelib==1.4.2
readability==0.2
redis==2.10.6
redis-py-cluster==1.3.4
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.5.0
selenium==3.9.0
service-identity==17.0.0
shellescape==3.4.1
six==1.11.0
sklearn==0.0
smart-open==1.5.6
snownlp==0.12.3
SQLAlchemy==1.2.4
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
Theano==1.0.1
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
